refactor(form): split FormErrorDiscovery (P12d)
Phase 12d file 5/8 (Login + Form HIGH cluster decoupling).

Why
- FormErrorDiscovery.ts was a 256-line god-module mixing two
  discovery layers (dynamic DOM scan + WK text probe), private
  helpers, types and constants in one file.
- Splitting per CLEAN_CODE.md per-function caps + the Phase 12d
  cluster-decoupling pattern proven by PR #339.

What
- Extracted ErrorDiscovery/ErrorDiscoveryTypes.ts: shared types
  (FormErrorKind, IFormError, IFormErrorScanResult, IRawDomItem,
  IEvalArg) + constants (NO_CLASS, ERROR_SELECTOR, NO_ERRORS).
- Extracted ErrorDiscovery/ErrorDiscoveryScan.ts: Layer 1
  dynamic DOM scan + private helpers (scanDomErrorsInBrowser,
  queryDomErrors, classifyByTag, buildSelector, toFormError,
  isKnownErrorText, isDedicatedErrorTag, filterVisible).
- Extracted ErrorDiscovery/ErrorDiscoveryProbe.ts: Layer 2 WK
  text probe + private probeWellKnownText helper.
- Rewrote FormErrorDiscovery.ts as a 28-LoC re-export barrel
  preserving all 4 public types + 2 functions byte-identically.

Guideline compliance
- before-commit-guidlines.md: all 23 husky gates expected GREEN
  (lint, lint:architecture, lint:canaries, tsc, eslint clean,
  39/39 targeted tests, docs-coverage allowlisted).
- eslint-rules-guidlines.md S3: zero eslint-disable headers
  added; the single-export prefer-default-export rule satisfied
  by re-exporting IFormErrorScanResult from the sibling Types
  module in both Scan and Probe files.
- eslint-rules-guidlines.md S5: no eslint.config.mjs changes;
  refactor commit is separable from any guardrail commit.
- comments-in-code-guidlines.md / j-doc-guidlines.md: BLUF
  JSDoc on every exported + inline arrow function with @param
  + @returns; no commented-out code.
- general-rules-guidlines.md P7 Mediator: both functions remain
  mediator handlers; LoginSteps still calls them through
  IElementMediator.discoverErrors.
- CLAUDE.md ZERO CSS in interaction code: structural CSS
  selectors are confined to ERROR_SELECTOR (parsing/extraction
  surface) - explicitly allowed per CLAUDE.md.

Tracker(s)
- Phase 12d
- PVTI_lAHOAPtF084BZbzwzgvauEI
